refactor(sdk): Group the attest modules under src/attest
Mirrors the popup module layout: bindings, context, and submission in
attest/index.ts, the canonical registry map in attest/registries.ts.
Public import paths are unchanged.
