+++ packages/zkpassport-sdk/tests/popup.test.ts
@@ -3,7 +3,7 @@ import type { Query } from "@zkpassport/utils"
 import { ZKPassport } from "../src/index"
 import { hydrateQueryBuilder } from "../src/popup/hydrate"
 import { openVerificationPopup, isPopupMessage } from "../src/popup"
-import { getAttestRegistry } from "../src/attest-registries"
+import { getAttestRegistry } from "../src/attest/registries"
 import type { QueryBuilder } from "../src/types"
 
 function buildOffline(build: (qb: QueryBuilder<"offline">) => unknown): Query {
